==> Block n°1/01_Kayak_request_api.ipynb <==
L'objectif de ce projet est de "Planifier un voyage avec Kayak" :

Kayak est un moteur de recherche qui aide l'utilisateur à planifier son voyage au meilleur prix.

Booking est un site internet qui référence des hébergements dans différents types de locaux allant de l'hôtel au gîte touristique (bed and breakfast) en passant par l'appartement

Nous allons créer une recommandation personnalisée pour l'utilisateur en fonction de certains critères météorologiques (température, vitesse du vent, taux d'humidité) en créant un score de ces paramètres qui permettra d'aider l'utilisateur à choisir sa destination pour son voyage.

Nous allons aussi utiliser le site de Booking pour recupérer les 25 meilleurs hotels des destinations que l'on possède qui est une liste de 35 villes.

PROJET:</br>
Nous travaillons pour l'équipe marketing qui s'est rendu compte que **70 % de leurs utilisateurs qui planifient un voyage aimeraient avoir plus d'informations sur la destination vers laquelle ils se rendent**</br>
Cette equipe souhaite créer une application qui recommandera les meilleurs endroits de planification des vacances de leur client. Cette application sur :

* Temps
* Hotels dans la region

Nous allons donc devoir :

* Utiliser OpenStreetMap pour recupérer les coordonnées des villes que nous souhaitons recommandés
* Manipuler OpenWeatherMap pour recupérer les données météorologiques de ces villes
* Scraper les données des meilleurs hotels de ces villes
* Créer un classement en fonction des conditions météorologiques 
* Constituer 2 Dataframes :
    * 1 pour les villes en fonction de la météo
    * 1 pour les meilleurs hotels de ces villes
* Exploiter AWS S3 pour stocker les données (qui est un Data lake) en utilisant boto3
* Extraire, transformer et charger les données nettoyées de S3 à une base de donnée AWS RDS en utilisant SQL Alchemy
* Créer 2 cartes : 
    * 1 qui décrit les 35 villes avec les conditions météo
    * 1 qui affiche pour les 5 meilleures villes ( qui ont le meilleur score) les meilleurs hotels associés 

J'ai donc decomposé ce projet en 5 parties:

* 1ere Partie : Recupération des coordonnées avec OpenStreetMap et les conditions météorologiques avec OpenWeatherMap des villes que nous avons ciblées
* 2eme Partie : Scraper les données des 25 meilleurs hotels en fonction de certains paramètres
* 3eme Partie : Formater les données pour obtenir 2 Dataframes (un pour les villes avec leur météo associé et un pour les 25 meilleurs hotels)
* 4eme Partie : Transferer les données dans un Data Lake, puis les charger dans une base de donnée type Data Warehouse
* 5eme Partie : Créer les 2 maps qui représentent TOP 5 des destinations et TOP 20 des hotels associés

import requests
import pandas as pd
import json
import os
from collections import Counter
import numpy as np

cities = ["Mont Saint Michel","St Malo","Bayeux","Le Havre","Rouen","Paris","Amiens","Lille","Strasbourg",
          "Chateau du Haut Koenigsbourg","Colmar","Eguisheim","Besancon","Dijon","Annecy","Grenoble","Lyon",
          "Gorges du Verdon","Bormes les Mimosas","Cassis","Marseille","Aix en Provence","Avignon","Uzes",
          "Nimes","Aigues Mortes","Saintes Maries de la mer","Collioure","Carcassonne","Ariege","Toulouse",
          "Montauban","Biarritz","Bayonne","La Rochelle"]

# si le fichier est dans un .txt:
"""
def extract_txt_to_list(path_to_file, file):
    import os
    p = path_to_file
    os.chdir(p)
    s = []
    with open(file, 'r') as f:
        for line in f:
            if line[0] == '[':
                s.append(line[2:-3])
            elif line[-1] == ']':
                s.append(line[1:-2])
            else:
                s.append(line[1:-3])
                
    return s

path = r'C:\Users\david\JEDHA BOOTCAMP\FULLSTACK\10_PROJECT\Kayak'
file = 'cities_name.txt'

cities = extract_txt_to_list(path, file)
"""

Nous allons commencer à recupérer les coordonnées des villes avec l'API OpenStreetMap et les données météorologiques avec OpenWeatherMap

Voici une fonction qui permet d'afficher certaines caractéristiques (comme les coordonnées GPS, la Température, le ressenti, le taux d'humidité, ect) pour une ville donnée

En requetant l'API (et en appliquant .json() sur le requests.get()), on se trouve avec un .json qui contient comme clés :</br>
 ['coord', 'weather', 'base', 'main', 'visibility', 'wind', 'rain', 'clouds', 'dt', 'sys', 'timezone', 'id', 'name', 'cod']

 On peut donc utiliser certains clés qui nous donnerons les valeurs associées pour avoir un rapide aperçu de conditions météo et des coordonnées de la ville ciblée en fonctions de son nom

# Test 

city_name = 'Bayeux,FR'
API_key= "2f2685efc16594e365e34e00af802902"

def get_weather(API_key, city, return_=False):
    url = f"http://api.openweathermap.org/data/2.5/weather?q={city}&appid={API_key}"
    r = requests.get(url).json()
    print(f"Ville :\n {r['name']}\n")
    print(f"Coordonnées GPS :\n Longitude : {r['coord']['lon']}\n Latitude : {r['coord']['lat']}\n")
    print(f"Conditions Météorologiques :\n Temperature(C°) : {int(r['main']['temp'] - 273.15)} \n Ressenti(C°) : {int(r['main']['feels_like'] - 273.15)} \n Humidité : {r['main']['humidity']}")
    print(f" Météo : {r['weather'][0]['main']}")
    if return_:
        return r

get_weather(API_key, city_name)

Nous créons une fonction qui permet de rechercher les coordonnées GPS en requetant OpenStreetMap et la librairie requests:


def create_coords(cities):
    s = []
    url = 'https://nominatim.openstreetmap.org/search'
    cols = ['city_id' , 'name', 'latitude', 'longitude']
    n = len(cities)
    for i in range(n):
        params = {"city": cities[i], "country": "France", "format": "json"}
        r = requests.get(url, params=params).json()
        s.append(r)
        
    for i in range(len(s)):
        if s[i] == []:
            print(cities[i], 'not in the list')
            
    cities_id = [i for i in range(n) if s[i] != []]
    cities_name = [city for i,city in enumerate(cities) if s[i] != []]
    cities_lat = [s[i][0]['lat'] for i in range(n) if s[i] != []]
    cities_lon = [s[i][0]['lon'] for i in range(n) if s[i] != []]
    
    rows = [cities_id, cities_name, cities_lat, cities_lon]  
    df = pd.DataFrame({cols[0] : cities_id, cols[1] : cities_name, cols[2] : cities_lat, cols[3] : cities_lon})
    
    return df

Il y a 2 villes qui n'ont pas de coordonnées : 
* Gorges du Verdon (car ce n'est pas une ville a proprement parlé, c'est plutôt un ensemble de Gorges ou de Canyons)
* Ariege (car ce n'est pas une ville mais un département)

Nous allons donc récupérer les données autrement : 

il y a 2 possibilités :</br>
* soit on recupère directement les données GPS avec Google Map en ciblant ces 2 endroits
* soit on change(pour Ariège) l'url de request par 'county': 'https://nominatim.openstreetmap.org/search?county=' au lieu de 'city'

La première solution est préférable car la seconde demande de faire plusieurs manipulations (d'abord recupérer les coordonnées pour Ariège en ciblant 'county' au lieu de 'city' puis soit de recupérer les coordonnées de Gorges du Verdon avec Google Maps soit ceux d'une ville proche)

df = create_coords(cities)
df.head()

On crée un DataFrame contenant les noms des villes, ainsi que leurs coordonnées respectives

df_temp_ = pd.DataFrame({'city_id' : [17,29], 'name' : ['Gorges du Verdon', 'Ariege'], 'latitude' : [43.7496562, 42.9455368],
                         'longitude' : [6.3285616, 1.4065544]})

df = pd.concat([df,df_temp_], axis=0).sort_values(by='city_id')
df.head()

df.shape

df.to_csv('kayak2_df.csv', index=None)

df = pd.read_csv('kayak2_df.csv')
df.head()

 J'ai choisi de prendre les données concernant les conditions météorologiques sur une période de 7 jours :

 Si un client souhaite prendre une semaine de vacances, cela lui donnera un apercu du temps sur cette période (sachant que plus le temps consacré au vacances est long, plus la météo est importante dans son critère de destination car il serait dommage d'avoir du beau temps le 1er jour et tout le reste de la semaine, le temps serait dégradé).

 Il faut quand même prendre en considération que plus le temps passe, moins la météo est fiable sur ses prévisions : dans ce cas, il serait judicieux d'appliquer un coefficient qui diminuera la valeur en fonction de la durée : plus on s'eloigne dans le temps, moins les valeurs auront du poids.

Voici une fonction qui permet de récuperer les conditions météorologiques en utilisant l'API de OpenWeatherMap :</br>
d'abord nous requetons l'API pour chaque ville : nous obtenons un objet .json avec plusieurs clés :

 ['lat', 'lon', 'timezone', 'timezone_offset', 'hourly', 'daily']

 Il y en a plusieurs qui nous interesse :
 * Tout d'abord, les coordonnées GPS qui sont `lat` et `lon`
 * Ensuite, on peut même connaitre les conditions météo par heure avec `hourly` (ce qui peut être interessant mais nous ne le selectionnerons pas pour cette étude)
 * Enfin, le plus important est la clé `daily` qui recense tous les conditions météo par journée sur une période de 7 jours pour ma version d'API

 Dans cette clé `daily`, nous commencerons au lendemain (r['daily'][1:]) car nous ne savons pas quand le client fera sa recherche.

 Plusieurs paramètres nous interesse comme le ressenti (j'ai volontairement remplacé la température par le ressenti car la température qui nous impacte est celle ressentie), le taux d'humidité, la probabilité de pleuvoir, la vitesse du vent (plus il y a de vent, et plus cela peut être désagréable pour le client), la pression atmosphérique, et la nature du ciel un jour donné (dégagé, couvert, pluvieux, ect..)

 On peut donc récuperer chaque critère pour les 7 jours à venir sous forme de liste et j'ai aussi fait la moyenne (le mode pour la nature du ciel) pour ces critères qui me donnent non pas une liste, mais un nombre.

On crée une variable `score` en mettant des coefficiants que le client peut modifier à sa guise en fonction de son envie (préfère t'il plus pénaliser le taux d'humidité ou la vitesse du vent, de combien veut il valoriser la température, ect..) ce qui affectera les moyennes (mode) précédemment crées.

Cette fonction renvoie un DataFrame avec l'ensemble de ces nouveaux paramètres ajoutés.

def get_weather_next_seven_days(df, API_key, coef_temp=2, coef_proba_rain=0.6, coef_himidity=0.2, coef_wind=0.9):
    
    from datetime import datetime
    now = datetime.now()
    temperature, rains, humidity, winds, pressures, descs = [], [], [], [], [], []
    temp_df = pd.DataFrame()
    url = "https://api.openweathermap.org/data/2.5/onecall"
    
    for i in range(len(df)):
        
        params = {
            "lat": df.loc[i,'latitude'],
            "lon": df.loc[i,'longitude'],
            "exclude": "current, minutely, hourly", # Exclusion de certains parametres de  l'API: current, minutely and hourly
            "units": "metric", 
            "appid": "2f2685efc16594e365e34e00af802902", # API key
        }
        
        r = requests.get(url,params).json()
        
        seven_next_days = r['daily'][1:] 
        
        feels = [int(weather['feels_like']['day']) for weather in seven_next_days] # temperature pour les 7 prochains jours
        rain = [int(weather['pop'] * 100) for weather in seven_next_days]          # pourcentage de pluie pour les 7 prochains jours
        humid = [weather['humidity'] for weather in seven_next_days]               # pourcentage d'humidité pour les 7 prochains jours
        wind = [weather['wind_speed'] for weather in seven_next_days]              # vitesse du vent
        pressure = [weather['pressure'] for weather in seven_next_days]            # pression atmospherique
        desc = [weather['weather'][0]['main'] for weather in seven_next_days]      # description des conditions meteos
        
        temperature.append(feels)
        rains.append(rain)    
        humidity.append(humid)
        winds.append(wind)
        pressures.append(pressure)
        descs.append(desc)
        
    temp_df['days'] = [[i for i in range(1, 8)] for row in range(len(df))]
    
    temp_df['temperature'] = temperature
    temp_df['mean_temp'] = temp_df['temperature'].apply(lambda x: int(np.array(x).mean()))
    temp_df['proba_rain'] = rains
    temp_df['mean_rain'] = temp_df['proba_rain'].apply(lambda x: np.round(np.array(x).mean()),2)
    temp_df['humidity'] = humidity
    temp_df['mean_humidity'] = temp_df['humidity'].apply(lambda x: np.round(np.array(x).mean()),2)
    temp_df['wind_speed'] = winds
    temp_df['mean_wind_speed'] = temp_df['wind_speed'].apply(lambda x: np.round(np.array(x).mean()),2)
    temp_df['pressure'] = pressures
    temp_df['mean_pressure'] = temp_df['pressure'].apply(lambda x: int(np.array(x).mean()))
    temp_df['weather'] = descs
    temp_df['mode_weather'] = temp_df['weather'].apply(lambda x:pd.Series(x).mode())
    
    mean_temp = [sum(el)/7 for el in temp_df['temperature']]
    mean_prob_rain = [sum(el)/7 for el in temp_df['proba_rain']]
    mean_humid = [sum(el)/7 for el in temp_df['humidity']]
    mean_wind = [sum(el)/7 for el in temp_df['wind_speed']]
    
    score = np.round((np.array(mean_temp)*coef_temp + np.array(mean_prob_rain)*coef_proba_rain + np.array(mean_humid)*coef_himidity + np.array(mean_wind)*coef_wind) / 3,2) 
    
    temp_df['score'] = score  
    temp_df['date'] = now.strftime("%m/%d/%Y, %H:%M:%S")
    
    return pd.concat([df,temp_df],axis=1)

df_ = get_weather_next_seven_days(df, API_key)
df_.head()

df_.columns

Voici une fonction qui énumère les différentes conditions météorologiques de la nature du ciel : on crée pour chaque catégorie de nature du ciel (dégagé, couvert, ect..) une colonne qui recense le nombre de fois ou cette nature apparait : ca permettra de crée un score qui pénalisera les plus mauvaises conditions et valorisera les meilleures.

def mode_weather(df):
    rain, clouds, clear = [], [], []
    for weather in df_['weather']:
        rain_count, clouds_count, clear_count = 0,0,0
        for el in weather:
            if el == 'Rain':
                rain_count += 1
            if el == 'Clouds':
                clouds_count += 1
            if el == 'Clear':
                clear_count += 1
        rain.append(rain_count)
        clouds.append(clouds_count)
        clear.append(clear_count)
    df['mode_rain'] = rain
    df['mode_clouds'] = clouds
    df['mode_clear'] = clear
    
    return df

df2 = df_.copy()
df2 = mode_weather(df2)
df2.columns

Dans cette fonction, on crée plusieurs `rank`, un par critères météo et on en ressort un **super rank** qui est une combinaison linéaire de tous les ranks (modifiable par le client encore une fois)

def rank_df(df, coef_rain=0.2, coef_clouds=0.5, coef_clear=1, coef_weather=0.35, coef_temp=2, coef_himidity=0.2, coef_wind=0.9, coef_pressure=0.004):    
    # Calcul du rang Meteo : 0.2 * Pluvieux + 0.5 * Nuageux + 1 * Ciel_dégagé : on penalise le mauvais temps
    df['rank_weather'] = df['mode_rain']*coef_rain + df['mode_clouds']*coef_clouds + df['mode_clear']*coef_clear

    df['rank_weather'] = df['rank_weather'].rank(method='max').astype(int)
    df['rank_rain'] = df['mean_rain'].rank(method='min').astype(int)
    df['rank_temp'] = df['mean_temp'].rank(method='max').astype(int)
    df['rank_humi'] = df['mean_humidity'].rank(method='min').astype(int)
    df['rank_wind'] = df['mean_wind_speed'].rank(method='min').astype(int)
    df['rank_pres'] = df['mean_pressure'].rank(method='max').astype(int)


    # Calcul final du rang : 0.5 * rang_pluie + 0.35 * rang_meteo + 0.3 * rang_temperature + 0.9 * rang_humidité + 0.2 * rang_pression_atmospherique / nombre de rangs
    df['rank_'] = (df['rank_rain']*coef_rain + df['rank_weather']*coef_weather + df['rank_temp']*coef_temp + df['rank_humi']*coef_himidity + df['rank_wind']*coef_wind + df['rank_pres']*coef_pressure)/6
    df = df.sort_values(by='rank_').drop(df.columns[-7:], axis=1)
    df['rank'] = [i+1 for i in range(len(df))]
    df = df.reset_index().drop(['index'],axis=1)
    return df

df3 = df2.copy()
df3 = rank_df(df2)
df3.columns

df3.head()

On vérifie les 5 meilleures destinations et on associera le resultat à une variable

df3['name'][:5]

best_five_cities = df3['name'][:5].values.tolist()
best_five_cities

df3.to_csv('./src/kayak2_final.csv', index=False)

df3 = pd.read_csv('./src/kayak2_final.csv')
df3.head()
