==> Block n°1/03_Kayak_formating_dfs.ipynb <==
### KAYAK : Formatage des 2 DataFrames (Weather & Hotel)

Dans ce notebook, nous allons preparer chaque DataFrame pour qu'on puisse faire de la Viz plus tard.

Dans un premier temps, nous allons remplacer les espaces par des '-', ce qui evitera d'avoir des problèmes de format quand nous ferons de la visualisation de données.

Ensuite, nous allons concatener les DataFrame qui recense les conditions météo des villes selectionnées et celui qui recense les meilleurs hotels.

import requests
import pandas as pd
import json
import os
from collections import Counter
import numpy as np

cities = ["Mont Saint Michel","St Malo","Bayeux","Le Havre","Rouen","Paris","Amiens","Lille","Strasbourg",
          "Chateau du Haut Koenigsbourg","Colmar","Eguisheim","Besancon","Dijon","Annecy","Grenoble","Lyon",
          "Gorges du Verdon","Bormes les Mimosas","Cassis","Marseille","Aix en Provence","Avignon","Uzes",
          "Nimes","Aigues Mortes","Saintes Maries de la mer","Collioure","Carcassonne","Ariege","Toulouse",
          "Montauban","Biarritz","Bayonne","La Rochelle"]

On récupère le DataFrame `filename` qui est le fichier .json qui est le resultat du scraping des 25 meilleurs hotels par destination et qui a été converti en .csv 

# df = pd.read_json(filename)
df = pd.read_csv('./src/filename.csv')

df.shape

temp = []
for el in cities:
    temp.append(el.replace(' ','-'))

result = []
for i,city in enumerate(temp):
    if city in df.loc[i,'description']:
        result.append((i,city))

result

Nous intégrons le nom des villes dans une colonne `Ville`

cities = [el.replace(" ","-") for el in cities]
cities = [((city + ' ') * 25).split() for city in cities]
cities = [c for city in cities for c in city]

len(cities)

df['Ville'] = cities
#df.drop(['location'],axis=1,inplace=True)

Test sur le nom d'une ville

df[df['Ville'] == 'Mont-Saint-Michel']

df.to_csv('./src/best_hotels.csv', index=False)

df.groupby("Ville").agg({"title": lambda x: list(x)}).head()

df_villes = df.groupby("Ville").agg(list)

kayak = pd.read_csv("./src/kayak2_final.csv")
kayak['name'] = kayak['name'].apply(lambda x:x.replace(' ','-'))
kayak

df_final = kayak.merge(df_villes, left_on="name", right_on="Ville")

On concatène les 2 DataFrames

df_final.sort_values('rank').head()

On crée un DataFrame pour les 5 meilleures destinations

best_five_cities = ['Rouen','St-Malo','Bayeux','Chateau-du-Haut-Koenigsbourg','Strasbourg']

df_flatten_final = df[df['Ville'].isin(best_five_cities)].merge(kayak,left_on="Ville", right_on="name").sort_values(['rank','score_x'])
df_flatten_final.head()

df_flatten_final.to_csv('./src/df_flatten_final2.csv', index=False)

### Formatage du dataset kayak2_final pour la Viz

On va devoir applatir toutes les listes qui sont dans les colonnes de 'kayak2_final' pour pouvoir faire une Viz par date, pour cela, on va créer des foncitons appropriées.

k = kayak.copy()

def convert_string_to_list(df, cols_list):
    import ast
    
    for col in cols_list:
        df[col] = df[col].apply(lambda x: ast.literal_eval(x))       
    df = df.explode(list_strings).reset_index().drop(['index'], axis=1)
    
#    for el in df.iloc[0].index:
#        if type(df[el][0]) == str:
#            df[col] = df[col].apply(lambda x: ast.literal_eval(x))
        
    return df

list_strings = ['days', 'temperature', 'proba_rain', 'humidity', 'wind_speed',  'pressure', 'weather']

flatten_weather = convert_string_to_list(k, list_strings)
flatten_weather

flatten_weather.to_csv('./src/flatten_weather.csv', index=False)

def unnesting(df, explode, axis=0):
    if axis==1:
        df1 = pd.concat([df[x].explode() for x in explode], axis=1)
        return df1.join(df.drop(explode, axis=1), how='left')
    else :
        df1 = pd.concat([
                         pd.DataFrame(df[x].tolist(), index=df.index).add_prefix(x) for x in explode], axis=1)
        return df1.join(df.drop(explode, axis=1), how='left')

unnesting(k, list_strings, axis=0).head(1)

def formating_hotels(df_hotels):
    df_hotels_full = df_hotels.sort_values('score_weather')
    df_hotels_full.reset_index(inplace=True, drop=True)
    df_hotels_full = df_hotels_full.apply(pd.Series.explode)
    df_hotels_full['description'] = df_hotels_full['description'].replace("\\n", "", regex=True)
    df_hotels_full['score'] = df_hotels_full['score'].replace(",", ".", regex=True)
    df_hotels_full[['lat_hotels', 'lon_hotels', 'score']] = df_hotels_full[['lat_hotels', 'lon_hotels', 'score']].apply(pd.to_numeric)
    return df_hotels_full
