==> Block n°1/04_Kayak_SQL_S3.ipynb <==
### KAYAK

import requests
import pandas as pd
import json
import os
from collections import Counter
import numpy as np

cities = ["Mont Saint Michel","St Malo","Bayeux","Le Havre","Rouen","Paris","Amiens","Lille","Strasbourg",
          "Chateau du Haut Koenigsbourg","Colmar","Eguisheim","Besancon","Dijon","Annecy","Grenoble","Lyon",
          "Gorges du Verdon","Bormes les Mimosas","Cassis","Marseille","Aix en Provence","Avignon","Uzes",
          "Nimes","Aigues Mortes","Saintes Maries de la mer","Collioure","Carcassonne","Ariege","Toulouse",
          "Montauban","Biarritz","Bayonne","La Rochelle"]

### SQL 

from sqlalchemy import create_engine, text
from sqlalchemy.orm import sessionmaker

DBHOST = "database-1.coculd4a9fzq.eu-west-3.rds.amazonaws.com"
DBUSER = "DataScientist"
DBPASS = "XXXXXXXXXXXXXXXXXX"
DBNAME = "postgres"
PORT = 5432

engine = create_engine(f"postgresql+psycopg2://{DBUSER}:{DBPASS}@{DBHOST}/{DBNAME}", echo=True)
Session = sessionmaker(bind=engine)
session = Session()

# On convertie les dataframes en table SQL
df_flatten_final.to_sql('weather', con=engine, if_exists='replace', index=False)

# Test des data avec une query
query = text("SELECT * FROM weather LIMIT 21")
pd.read_sql(query, engine)

### S3 Buckets

import boto3
YOUR_ACCESS_KEY = "XXXXXXXXXXXXXXXXXXXX"
YOUR_SECRET_KEY = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"

session = boto3.Session(aws_access_key_id="YOUR_ACCESS_KEY", 
                        aws_secret_access_key="YOUR_SECRET_KEY")

s3 = session.resource("s3")

kayak_bucket = s3.create_bucket(Bucket="kayak_project")

csv_weather = kayak.to_csv()
csv_hotels = df_total.to_csv()
csv_top5 = df_flatten_final.to_csv()

put_object = bucket.put_object(Key="weather.csv", Body=csv_weather)
put_object = bucket.put_object(Key="hotels.csv", Body=csv_hotels)
put_object = bucket.put_object(Key="top5.csv", Body=csv_top5)

kayak = pd.read_csv("kayak_final.csv")
df_total = pd.read_csv("best_hotels.csv")
df_flatten_final = pd.read_csv('df_flatten_final.csv')

df_weather = kayak.copy()
df_hotels = df_total.copy()
df_top5 = df_flatten_final.copy()
